+++ src/main/kotlin/net/trilleo/Hex.kt
@@ -4,6 +4,7 @@ import net.fabricmc.api.ClientModInitializer
 import net.minecraft.resources.Identifier
 import net.trilleo.feature.Features
 import net.trilleo.keybind.KeybindsFeature
+import net.trilleo.update.UpdateFeature
 import org.slf4j.LoggerFactory
 
 /**
@@ -17,6 +18,7 @@ object Hex : ClientModInitializer {
 
 	override fun onInitializeClient() {
 		Features.register(KeybindsFeature)
+		Features.register(UpdateFeature)
 		Features.bootstrap()
 
 		LOGGER.info("Hex initialized")